Seperated the code for each method
